download eta; like/don't like buttons
git-svn-id: http://aptoide.org/repo/aptoide-client@1710 8d55eb69-6847-4b75-98a7-935b3913b0ec

+++ unstable/Aptoidev5/trunk/v5engine/src/main/java/cm/aptoide/ptdev/adapters/OngoingAdapter.java
@@ -68,7 +68,7 @@ public class OngoingAdapter extends ArrayAdapter<Download> {
         View downloadsLayout = v.findViewById(R.id.download_details_layout);
         downloadsLayout.setVisibility(View.VISIBLE);
         ((TextView) downloadsLayout.findViewById(R.id.speed)).setText(Utils.formatBits((long) download.getSpeed())+"/s");
-        ((TextView) downloadsLayout.findViewById(R.id.eta)).setText(Utils.formatEta((long) download.getTimeLeft(), context.getString(R.string.time_left)));
+        ((TextView) downloadsLayout.findViewById(R.id.eta)).setText(Utils.formatEta(download.getTimeLeft(), context.getString(R.string.time_left)));
         ((TextView) downloadsLayout.findViewById(R.id.progress)).setText(download.getProgress()+"%");
